In Makefile, name output after build directory
This feature causes the output files to take on the name of the
directory that this content lives in. We would expect direct
clones to generate output named presentation-template.pdf.

However, and this is the cool part, when the user grabs a copy
of this repo and renames it (i.e., for a specific presentation)
then the output files naming is automatically updated and the
output files match the new repo name. For example, if I created
a new copy of this repo for my alife-2018 presentation (i.e.,
setting up a new repo called alife-2018 and renaming the repo
folder alife-2018), my output would automatically be named
alife-2018.pdf when I compile it.

I am unsure of the cross-platform compatability of the Makefile
magic I used to grab the name of the containing directory, so
I added a follow-up that should define BUILD_DIR as "presentation"
if BUILD_DIR isn't successfully defined by the Makefile magic.